+++ backend/windmill-api/src/oauth2.rs
@@ -39,7 +39,7 @@ use crate::{
     workspaces::WorkspaceSettings,
     BaseUrl,
 };
-use windmill_common::error::{self, to_anyhow, Error, Result};
+use windmill_common::error::{self, to_anyhow, Error};
 use windmill_common::oauth2::*;
 
 use windmill_queue::JobPayload;
@@ -53,7 +53,6 @@ pub fn global_service() -> Router {
         .route("/connect/:client", get(connect))
         .route("/connect_callback/:client", post(connect_callback))
         .route("/connect_slack", get(connect_slack))
-        .route("/connect_slack_callback", post(connect_slack_callback))
         .route(
             "/slack_command",
             post(slack_command).route_layer(axum::middleware::from_extractor::<SlackSig>()),
@@ -66,16 +65,17 @@ pub fn workspaced_service() -> Router {
     Router::new()
         .route("/disconnect/:id", post(disconnect))
         .route("/disconnect_slack", post(disconnect_slack))
-        .route("/set_workspace_slack", post(set_workspace_slack))
         .route("/create_account", post(create_account))
         .route("/delete_account/:id", post(delete_account))
         .route("/refresh_token/:id", post(refresh_token))
+        .route("/connect_slack_callback", post(connect_slack_callback))
 }
 
 pub struct ClientWithScopes {
     client: OClient,
     scopes: Vec<String>,
     extra_params: Option<HashMap<String, String>>,
+    allowed_domains: Option<Vec<String>>,
 }
 
 pub type BasicClientsMap = HashMap<String, ClientWithScopes>;
@@ -92,6 +92,7 @@ pub struct OAuthConfig {
 pub struct OAuthClient {
     id: String,
     secret: String,
+    allowed_domains: Option<Vec<String>>,
 }
 pub struct AllClients {
     pub logins: BasicClientsMap,
@@ -135,19 +136,16 @@ pub async fn build_oauth_clients(base_url: &str) -> anyhow::Result<AllClients> {
         .map(|(k, v)| {
             let scopes = v.scopes.clone();
             let extra_params = v.extra_params.clone();
-            let named_client = build_basic_client(
-                k.clone(),
-                v,
-                oauths.get(&k).unwrap().clone(),
-                true,
-                base_url,
-            );
+            let client_params = oauths.get(&k).unwrap().clone();
+            let named_client =
+                build_basic_client(k.clone(), v, client_params.clone(), true, base_url, None);
             (
                 named_client.0,
                 ClientWithScopes {
                     client: named_client.1,
                     scopes: scopes.unwrap_or(vec![]),
                     extra_params,
+                    allowed_domains: client_params.allowed_domains,
                 },
             )
         })
@@ -166,6 +164,7 @@ pub async fn build_oauth_clients(base_url: &str) -> anyhow::Result<AllClients> {
                 oauths.get(&k).unwrap().clone(),
                 false,
                 base_url,
+                None,
             );
             (
                 named_client.0,
@@ -173,6 +172,7 @@ pub async fn build_oauth_clients(base_url: &str) -> anyhow::Result<AllClients> {
                     client: named_client.1,
                     scopes: scopes.unwrap_or(vec![]),
                     extra_params,
+                    allowed_domains: None,
                 },
             )
         })
@@ -190,6 +190,7 @@ pub async fn build_oauth_clients(base_url: &str) -> anyhow::Result<AllClients> {
             v.clone(),
             false,
             base_url,
+            Some(format!("{base_url}/oauth/callback_slack")),
         )
         .1
     });
@@ -203,12 +204,15 @@ pub fn build_basic_client(
     client_params: OAuthClient,
     login: bool,
     base_url: &str,
+    override_callback: Option<String>,
 ) -> (String, OClient) {
     let auth_url = Url::parse(&config.auth_url).expect("Invalid authorization endpoint URL");
     let token_url = Url::parse(&config.token_url).expect("Invalid token endpoint URL");
 
     let redirect_url = if login {
         format!("{base_url}/user/login_callback/{name}")
+    } else if let Some(callback) = override_callback {
+        callback
     } else {
         format!("{base_url}/oauth/callback/{name}")
     };
@@ -555,11 +559,14 @@ async fn connect_callback(
 }
 
 async fn connect_slack_callback(
+    Path(w_id): Path<String>,
+    authed: Authed,
     cookies: Cookies,
     Json(callback): Json<OAuthCallback>,
+    Extension(user_db): Extension<UserDB>,
     Extension(clients): Extension<Arc<AllClients>>,
     Extension(http_client): Extension<Client>,
-) -> error::JsonResult<SlackTokenResponse> {
+) -> error::Result<String> {
     let client = clients
         .slack
         .as_ref()
@@ -568,22 +575,13 @@ async fn connect_slack_callback(
     let token =
         exchange_code::<SlackTokenResponse>(callback, &cookies, client, &http_client).await?;
 
-    Ok(Json(token))
-}
-
-async fn set_workspace_slack(
-    Path(w_id): Path<String>,
-    Json(token): Json<SlackTokenResponse>,
-    Extension(user_db): Extension<UserDB>,
-    authed: Authed,
-) -> Result<String> {
     let mut tx = user_db.begin(&authed).await?;
 
     sqlx::query!(
         "INSERT INTO workspace_settings
-            (workspace_id, slack_team_id, slack_name)
-            VALUES ($1, $2, $3) ON CONFLICT (workspace_id) DO UPDATE SET slack_team_id = $2, \
-         slack_name = $3",
+                (workspace_id, slack_team_id, slack_name)
+                VALUES ($1, $2, $3) ON CONFLICT (workspace_id) DO UPDATE SET slack_team_id = $2, \
+             slack_name = $3",
         &w_id,
         token.team_id,
         token.team_name
@@ -592,13 +590,13 @@ async fn set_workspace_slack(
     .await?;
     sqlx::query!(
         "INSERT INTO group_
-            (workspace_id, name, summary)
-            VALUES ($1, $2, $3) ON CONFLICT DO NOTHING",
+                (workspace_id, name, summary)
+                VALUES ($1, $2, $3) ON CONFLICT DO NOTHING",
         &w_id,
         "slack",
         "The group that runs the script triggered by the slack /windmill command.
-                     Share scripts to this group to make them executable from slack and add
-                     members to this group to let them manage the slack related owner space."
+                         Share scripts to this group to make them executable from slack and add
+                         members to this group to let them manage the slack related owner space."
     )
     .execute(&mut tx)
     .await?;
@@ -728,18 +726,25 @@ async fn login_callback(
     Extension(http_client): Extension<Client>,
     Extension(is_secure): Extension<Arc<IsSecure>>,
 ) -> error::Result<String> {
-    let client = (&clients
+    let client_w_config = &clients
         .logins
         .get(&client_name)
-        .ok_or_else(|| error::Error::BadRequest("invalid client".to_string()))?
-        .client)
-        .to_owned();
+        .ok_or_else(|| error::Error::BadRequest("invalid client".to_string()))?;
+    let client = client_w_config.client.to_owned();
     let token_res = exchange_code::<TokenResponse>(callback, &cookies, client, &http_client).await;
 
     if let Ok(token) = token_res {
         let token = &token.access_token.to_string();
 
         let email = get_email(&http_client, &client_name, token).await?;
+
+        if let Some(domains) = &client_w_config.allowed_domains {
+            if !domains.iter().any(|d| email.ends_with(d)) {
+                return Err(error::Error::BadRequest(format!(
+                    "domain is not in the list of allowed domains: {email}, allowed: {domains:#?}",
+                )));
+            }
+        }
 
         let mut tx = db.begin().await?;