+++ frontend/src/lib/components/InviteGlobalUser.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
 	import { sendUserToast } from '$lib/utils'
-	import Switch from './Switch.svelte'
 	import type Modal from './Modal.svelte'
 	import { createEventDispatcher } from 'svelte'
 	import { UserService } from '$lib/gen'
 	import { Button } from './common'
+	import Toggle from './Toggle.svelte'
 
 	const dispatch = createEventDispatcher()
 
@@ -46,7 +46,7 @@
 <div class="flex flex-row">
 	<input on:keyup={handleKeyUp} placeholder="email" bind:value={email} />
 
-	<Switch class="ml-2" bind:checked={is_super_admin} horizontal={true} label={'admin: '} />
+	<Toggle class="mx-2" bind:checked={is_super_admin} options={{ right: 'admin' }} />
 	<input on:keyup={handleKeyUp} type="password" placeholder="password" bind:value={password} />
 	<input on:keyup={handleKeyUp} placeholder="name" bind:value={name} />
 	<input on:keyup={handleKeyUp} placeholder="company" bind:value={company} />