+++ frontend/src/lib/components/flows/content/FlowSettings.svelte
@@ -25,14 +25,7 @@
 
 		<svelte:fragment slot="content">
 			<TabContent value="metadata" class="p-4">
-				<Path bind:path={$flowStore.path} {initialPath} namePlaceholder="my_flow" kind="flow">
-					<div slot="ownerToolkit">
-						Flow permissions depend on their path. Select the group
-						<span class="font-mono"> all </span>
-						to share your flow, and <span class="font-mono">user</span> to keep it private.
-						<a href="https://docs.windmill.dev/docs/reference/namespaces">docs</a>
-					</div>
-				</Path>
+				<Path bind:path={$flowStore.path} {initialPath} namePlaceholder="my_flow" kind="flow" />
 
 				<label class="block my-4">
 					<span class="text-gray-700 text-sm">Summary <Required required={false} /></span>